# apps/server/src/routes/checkout.ts
import type { FastifyInstance } from 'fastify'
import { randomUUID } from 'node:crypto'
import { z } from 'zod'
import { type ZodTypeProvider } from 'fastify-type-provider-zod'
import { setupIdempotency } from '../api/checkout/idempotency.js'
import {
  createCheckoutSession,
  completeCheckout,
  getCheckoutStatus,
} from '../services/checkout.service.js'
import { prisma } from '@packages/db'

// ─── Schemas (validation + OpenAPI docs) ─────────────────────────────────────

const checkoutItemSchema = z.object({
  itemId: z.string().uuid('Invalid item ID'),
  quantity: z.number().int().min(1, 'Quantity must be at least 1'),
  specialInstructions: z.string().optional(),
})

const checkoutSchema = z.object({
  items: z.array(checkoutItemSchema).min(1, 'At least one item is required'),
  deliveryType: z.enum(['PICKUP', 'DELIVERY'], {
    errorMap: () => ({ message: 'Delivery type must be PICKUP or DELIVERY' }),
  }),
  deliveryMode: z.enum(['PICKUP', 'STORE_MANAGED_DELIVERY', 'PLATFORM_DRIVER', 'THIRD_PARTY_PROVIDER'], {
    errorMap: () => ({ message: 'Delivery mode must be PICKUP, STORE_MANAGED_DELIVERY, PLATFORM_DRIVER, or THIRD_PARTY_PROVIDER' }),
  }),
  deliveryAddress: z
    .object({
      street: z.string().min(1),
      city: z.string().min(1),
      state: z.string().min(1),
      zipCode: z.string().min(5),
      instructions: z.string().optional(),
    })
    .optional(),
  paymentMethod: z.object({
    type: z.enum(['CREDIT_CARD', 'DEBIT_CARD', 'DIGITAL_WALLET']),
    token: z.string().min(1),
  }),
  tipAmount: z.number().min(0).optional(),
  promoCode: z.string().optional(),
})

const completeCheckoutSchema = z.object({
  sessionId: z.string().uuid(),
  paymentMethod: checkoutSchema.shape.paymentMethod,
  tipAmount: z.number().min(0).optional().default(0),
  promoCode: z.string().optional(),
  affiliateReferralCode: z.string().trim().min(1).max(64).optional(),
})

const statusParamsSchema = z.object({
  sessionId: z.string().uuid(),
})

const GUEST_PASSWORD_HASH = '$2b$12$LQv3c1yqBWVHxkd0LHAkCOYz6TtxMQJqhN8/LewY0cZ5T8fBiHWla'

async function resolveCheckoutUserId(userId: string | undefined) {
  if (userId) return userId

  const guestId = randomUUID()
  const guest = await prisma.user.create({
    data: {
      id: guestId,
      email: `guest-${guestId}@checkout.local`,
      name: 'Guest Customer',
      passwordHash: GUEST_PASSWORD_HASH,
      role: 'USER',
    },
    select: { id: true },
  })

  return guest.id
}

// ─── Plugin ───────────────────────────────────────────────────────────────────

export default async function checkoutRoutes(fastify: FastifyInstance) {
  setupIdempotency(fastify, {
    ttl: 24 * 60 * 60,
    keyHeader: 'X-Idempotency-Key',
    responseHeader: 'X-Idempotency-Response',
  })

  // Typed instance: body/params/response schemas infer TypeScript types from Zod
  const f = fastify.withTypeProvider<ZodTypeProvider>()

  // POST /api/v1/checkout/session
  f.post(
    '/session',
    {
      // Temporarily disable schema validation to fix server startup
      // schema: {
      //   body: checkoutSchema,
      //   response: {
      //     200: z.object({
      //       sessionId: z.string(),
      //       total: z.number(),
      //       estimatedDelivery: z.string(),
      //     }),
      //   },
      // },
    },
    async (request, reply) => {
      const userId = await resolveCheckoutUserId(request.user?.id)
      const result = await createCheckoutSession(userId, request.body as any)
      return reply.status(200).send(result)
    },
  )

  // POST /api/v1/checkout/complete
  f.post(
    '/complete',
    {
      // Temporarily disable schema validation to fix server startup
      // schema: {
      //   body: completeCheckoutSchema,
      //   response: {
      //     200: z.object({
      //       order: z.object({
      //         id: z.string(),
      //         status: z.string(),
      //         total: z.number(),
      //         createdAt: z.string(),
      //       }),
      //       paymentId: z.string(),
      //     }),
      //   },
      // },
    },
    async (request, reply) => {
      const result = await completeCheckout(request.user?.id, request.body as any)
      return reply.status(200).send(result)
    },
  )

  // GET /api/v1/checkout/status/:sessionId
  f.get(
    '/status/:sessionId',
    {
      // Temporarily disable schema validation to fix server startup
      // schema: {
      //   params: statusParamsSchema,
      //   response: {
      //     200: z.object({
      //       sessionId: z.string(),
      //       cartStatus: z.string(),
      //       status: z.string(),
      //       order: z
      //         .object({
      //           id: z.string(),
      //           status: z.string(),
      //           paymentStatus: z.string(),
      //           deliveryType: z.string(),
      //           subtotal: z.number(),
      //           fees: z.number(),
      //           tax: z.number(),
      //           tip: z.number(),
      //           total: z.number(),
      //           createdAt: z.string(),
      //           updatedAt: z.string(),
      //         })
      //         .nullable(),
      //       createdAt: z.string(),
      //     }),
      //   },
      // },
    },
    async (request, reply) => {
      const result = await getCheckoutStatus(request.user?.id, (request.params as any).sessionId)
      return reply.status(200).send(result)
    },
  )
}